feat(yir): extract areachart
